refactor: moved TextDrawer and AbstractCairoDrawer to Cairo-specific module
