+++ src/igraph/drawing/shapes.py
@@ -23,7 +23,7 @@ from abc import ABCMeta, abstractmethod
 from math import atan2, copysign, cos, pi, sin
 import sys
 
-from igraph.drawing.baseclasses import AbstractCairoDrawer
+from igraph.drawing.cairo.base import AbstractCairoDrawer
 from igraph.drawing.matplotlib.utils import find_matplotlib
 from igraph.drawing.utils import Point
 from igraph.utils import consecutive_pairs
